requirements.txt
streamlit
psycopg2-binary
python-dotenv
black
isort
psycopg2